requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.9
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
flower==1.0.0
humanize==3.12.0
kombu==5.2.0
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
psycopg2-binary==2.8.6
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
